+++ differentfiles/colors.py
@@ -1,0 +1,9 @@
+black = (0, 0, 0)
+white = (255, 255, 255)
+light_gray = (255, 222, 173)
+dark_gray = (222, 184, 135)
+
+
+RED_HIGHLIGHT = (240, 50, 50, 150)
+
+GREEN_HIGHLIGHT = (0, 255, 0, 80)
