+++ differentfiles/pieces/__init__.py
@@ -1,0 +1,7 @@
+from .pawn import Pawn
+from .piece import Piece
+from .rook import Rook
+from .bishop import Bishop
+from .knight import Knight
+from .queen import Queen
+from .king import King
